Move header-only c10/util metaprogramming/type utils to torch/headeronly
Migrate six already-header-only c10/util headers into torch/headeronly/util/,
leaving the old c10 paths as one-line forwarders (fully backward compatible, no
consumer include changes):

  overloaded.h, IdWrapper.h, FunctionRef.h, Enumerate.h, C++17.h, ConstexprCrc.h

This is part of the wave moving c10 headers that need no .cpp/libtorch linking
into the header-only library. Symbols are defined canonically in
torch::headeronly and re-aliased into c10 (the bit_cast.h convention);
ConstexprCrc/C++17 live in the torch::headeronly::util/torch::headeronly::guts
sub-namespaces mirroring their original c10::util/c10::guts homes. IdWrapper's
hidden-friend hash_value/operator== travel with the class, so ADL keeps working
for derived id types (e.g. ConstexprCrc's crc64_t). C++17.h's guts::apply is
HIP-only and unchanged.

Test Plan:

Full build (confirms all libtorch consumers still compile):

```
pip install -e . --no-build-isolation
```

Header-only ABI check (compiles + runs without linking libtorch):

```
build/bin/test_aoti_abi_check --gtest_filter='TestOverloaded*:TestFunctionRef*:TestEnumerate*:TestIdWrapper*:TestConstexprCrc*'
```

Lint:

```
lintrunner -a <changed files>
```

This PR was authored with the assistance of Claude (Claude Code).

[ghstack-poisoned]
