+++ torch/headeronly/util/C++17.h
@@ -1,0 +1,42 @@
+#pragma once
+#ifndef C10_UTIL_CPP17_H_
+#define C10_UTIL_CPP17_H_
+
+#include <torch/headeronly/macros/Macros.h>
+#include <tuple>
+#include <utility>
+
+#if defined(__HIP__)
+
+HIDDEN_NAMESPACE_BEGIN(torch, headeronly, guts)
+
+// std::apply is not available in HIP device code because it lacks
+// __host__ __device__ annotations in the standard library.
+namespace detail {
+template <class F, class Tuple, std::size_t... INDEX>
+C10_HOST_DEVICE constexpr auto apply_impl(
+    F&& f,
+    Tuple&& t,
+    std::index_sequence<INDEX...>) {
+  return std::forward<F>(f)(std::get<INDEX>(std::forward<Tuple>(t))...);
+}
+} // namespace detail
+
+template <class F, class Tuple>
+C10_HOST_DEVICE constexpr auto apply(F&& f, Tuple&& t) {
+  return detail::apply_impl(
+      std::forward<F>(f),
+      std::forward<Tuple>(t),
+      std::make_index_sequence<
+          std::tuple_size<std::remove_reference_t<Tuple>>::value>{});
+}
+
+HIDDEN_NAMESPACE_END(torch, headeronly, guts)
+
+namespace c10::guts {
+using torch::headeronly::guts::apply;
+} // namespace c10::guts
+
+#endif
+
+#endif // C10_UTIL_CPP17_H_
